pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.0"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "loguru>=0.7.3",
    "celery>=5.6.0",
    "redis>=7.1.0",
    "types-pyyaml>=6.0.12.20250915",
    "motor>=3.7.1",
    "pymongo>=4.15.5",
    "rich>=14.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pylint>=3.3.8",
    "pytest>=8.4.2",
    "ruff>=0.13.1",
]


[tool.ruff]
target-version = "py311"
line-length = 150
output-format = "concise"
show-fixes = true
unsafe-fixes = false
preview = true
exclude = ["tutorials/", "tests/"]

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "B", # flake8-bugbear
    "UP", # pyupgrade,
    "I", # isort,
    "Q", # quote
    "W", # warning
]

unfixable = ["B"]

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"

[tool.pylint]
disable = [
    "too-many-instance-attributes",
    "too-few-public-methods",
    "too-many-return-statements",
    "too-many-locals",
    "too-many-arguments",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "import-outside-toplevel",
    "cyclic-import",
    "redefined-builtin",
    "broad-exception-caught",
    "broad-exception-raised",
    "super-init-not-called",
    "duplicate-code",
    "fixme",
    "logging-fstring-interpolation"
]
reports = false
max-branches=20
output-format = "colorized"
max-line-length = 150
class-const-naming-style = "PascalCase"
generated-members=["googleapiclient.*"]


[tool.pylint.DESIGN]
max-statements = 70

[tool.mypy]
no_implicit_optional = false
disallow_untyped_decorators = false
ignore_missing_imports = true
check_untyped_defs = true
warn_return_any = true
warn_no_return = false
warn_unreachable = true
strict = true
disallow_subclassing_any = false
strict_optional = true
plugins = ["pydantic.mypy"]
